:recycle: (pc-12740) extract logoutFromEduConnectIfAllowed function
